refactor(factory-ui): split BoardPage into board modules
BoardPage.tsx had grown to 1836 lines holding the page, both card
components, the column, the drag protocol, the candidate builders and
every query and mutation the board needs. Moves the non-markup parts
into domains/factory and leaves the page as its JSX tree (241 lines).

Board state now lives in useBoardState, with the composer's open lane
and the first-paint scroll positioning as private sub-hooks.

Drops three things that were already dead: the update mutation (created
but never called, so its isError could never fire), the deprecated
BoardPage export and its FactoryBoardPage wrapper, and the unused
fromStage parameter on moveItem. The stale-session filter that ran in
three places is now liveSessions, and the PR/Linear/approval prompts
that candidates and persisted items each built separately are now one
builder per run type.

No behaviour change: same 70 unit and 307 MSW tests pass as before.
