+++ mastracode/factory-ui/src/ui/domains/factory/components/IntakeColumnExtras.tsx
@@ -1,0 +1,51 @@
+import { Button } from '@mastra/playground-ui/components/Button';
+import { Txt } from '@mastra/playground-ui/components/Txt';
+
+import { useApiConfig } from '../../../../api/config';
+import type { useProjectIssuesQuery, useProjectPullRequestsQuery } from '../../../../hooks/useFactoryData';
+import type { useLinearIssuesQuery } from '../../../../hooks/useLinearData';
+import type { IntakeSource } from '../boardCandidates';
+import { connectLinear, isLinearReauthError } from '../services/linear';
+import { LoadMoreSentinel } from './LoadMoreSentinel';
+
+/**
+ * Intake column tail for the ACTIVE candidate feed: loading state, Linear
+ * reauth notice, and pagination. Only one feed is browsed at a time, so only
+ * its states render.
+ */
+export function IntakeColumnExtras({
+  source,
+  issues,
+  pulls,
+  linearIssues,
+}: {
+  source: IntakeSource | null;
+  issues: ReturnType<typeof useProjectIssuesQuery>;
+  pulls: ReturnType<typeof useProjectPullRequestsQuery>;
+  linearIssues: ReturnType<typeof useLinearIssuesQuery>;
+}) {
+  const { baseUrl } = useApiConfig();
+  if (source === null) return null;
+  const feed = source === 'github' ? issues : source === 'github-prs' ? pulls : linearIssues;
+
+  return (
+    <>
+      {source === 'linear' && linearIssues.isError && isLinearReauthError(linearIssues.error) && (
+        <div className="flex flex-col gap-2 p-1">
+          <Txt as="span" variant="ui-xs" className="text-icon3">
+            Linear authorization expired. Reconnect to keep syncing issues.
+          </Txt>
+          <Button size="xs" onClick={() => connectLinear(baseUrl)}>
+            Connect Linear
+          </Button>
+        </div>
+      )}
+      <LoadMoreSentinel
+        hasNextPage={Boolean(feed.hasNextPage)}
+        isFetchingNextPage={Boolean(feed.isFetchingNextPage)}
+        onLoadMore={() => void feed.fetchNextPage()}
+        label="Load more candidates"
+      />
+    </>
+  );
+}
